Simplified README, directory structure
